requirements.txt
fastapi==0.115.13
gradio==5.34.0
openai==1.88.0
Pillow==9.4.0
Pillow==11.2.1
pydantic==2.11.7
python-dotenv==1.1.0
Requests==2.32.4
striprtf==0.0.29
uvicorn==0.34.3
